Move collection sync logic to Signature class
Refactored FEP-8fcf collection synchronization logic by moving parse_collection_sync_header and xor_hex_strings from Http to Signature, and added compute_collection_digest to Signature. Updated Followers and Collection_Sync classes to use the new Signature methods for digest computation and header parsing. This centralizes collection sync logic and improves code organization.
